third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.cookies-blocked.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify Partitioned Popins cookie access when third-party cookie access blocked
  assert_equals: expected "ReadOnLoad:ThirdPartyNone-,ReadOnFetch:ThirdPartyNone-ThirdPartyNonePopin-,ReadOnDocument:ThirdPartyNone-ThirdPartyNonePopin-,ReadOnFetchAfterRSA:ThirdPartyNone-ThirdPartyNonePopin-ThirdPartyNonePopinAfterRSA-,ReadOnDocumentAfterRSA:ThirdPartyNone-ThirdPartyNonePopin-ThirdPartyNonePopinAfterRSA-,ReadInPopinIframe:ThirdPartyNone-,ReadInPopinIframeAfterRSA:FirstPartyNone-ThirdPartyNone-FirstPartyNoneAfterRSA-ThirdPartyNoneAfterRSA-" but got "ReadOnLoad:FirstPartyLax-FirstPartyNone-,ReadOnFetch:FirstPartyStrict-FirstPartyLax-FirstPartyNone-FirstPartyStrictPopin-FirstPartyLaxPopin-FirstPartyNonePopin-ThirdPartyStrictPopin-ThirdPartyLaxPopin-ThirdPartyNonePopin-,ReadOnDocument:FirstPartyStrict-FirstPartyLax-FirstPartyNone-FirstPartyStrictPopin-FirstPartyLaxPopin-FirstPartyNonePopin-ThirdPartyStrictPopin-ThirdPartyLaxPopin-ThirdPartyNonePopin-,ReadOnFetchAfterRSA:FirstPartyStrict-FirstPartyLax-FirstPartyNone-FirstPartyStrictPopin-FirstPartyLaxPopin-FirstPartyNonePopin-ThirdPartyStrictPopin-ThirdPartyLaxPopin-ThirdPartyNonePopin-FirstPartyStrictPopinAfterRSA-FirstPartyLaxPopinAfterRSA-FirstPartyNonePopinAfterRSA-ThirdPartyStrictPopinAfterRSA-ThirdPartyLaxPopinAfterRSA-ThirdPartyNonePopinAfterRSA-,ReadOnDocumentAfterRSA:FirstPartyStrict-FirstPartyLax-FirstPartyNone-FirstPartyStrictPopin-FirstPartyLaxPopin-FirstPartyNonePopin-ThirdPartyStrictPopin-ThirdPartyLaxPopin-ThirdPartyNonePopin-FirstPartyStrictPopinAfterRSA-FirstPartyLaxPopinAfterRSA-FirstPartyNonePopinAfterRSA-ThirdPartyStrictPopinAfterRSA-ThirdPartyLaxPopinAfterRSA-ThirdPartyNonePopinAfterRSA-,ReadInPopinIframe:ThirdPartyNone-,ReadInPopinIframeAfterRSA:FirstPartyNone-ThirdPartyNone-FirstPartyNoneAfterRSA-ThirdPartyNoneAfterRSA-"
Harness: the test ran to completion.
